+++ js-wrappers/src/main/java/org/jdiai/jswraper/JSWrapper.java
@@ -12,40 +12,43 @@ import static org.jdiai.jswraper.JSWrappersUtils.locatorsToBy;
 
 public class JSWrapper {
     public static Safe<WebDriver> DRIVER = new Safe<>(JSWrapper::chromeDriver);
+    public static WebDriver driver() {
+        return DRIVER.get();
+    }
 
     public static JSSmart $(String locator) {
-        return new JSSmart(DRIVER.get(), NAME_TO_LOCATOR.execute(locator));
+        return new JSSmart(driver(), NAME_TO_LOCATOR.execute(locator));
     }
     public static JSSmart $(String... locators) {
-        return new JSSmart(DRIVER.get(), locatorsToBy(locators));
+        return new JSSmart(driver(), locatorsToBy(locators));
     }
     public static <T> JSEntity<T> $(Class<T> cl, String locator) {
-        JSEntity<T> entity = new JSEntity<T>(DRIVER.get(), NAME_TO_LOCATOR.execute(locator)).initClass(cl);
-        entity.driver.setBuilder(new JSBuilder(DRIVER.get(), new SmartBuilderActions()));
+        JSEntity<T> entity = new JSEntity<T>(driver(), NAME_TO_LOCATOR.execute(locator)).initClass(cl);
+        entity.driver.setBuilder(new JSBuilder(driver(), new SmartBuilderActions()));
         return entity;
     }
     public static <T> JSEntity<T> $(Class<T> cl, String... locators) {
-        JSEntity<T> entity = new JSEntity<T>(DRIVER.get(), locatorsToBy(locators)).initClass(cl);
-        entity.driver.setBuilder(new JSBuilder(DRIVER.get(), new SmartBuilderActions()));
+        JSEntity<T> entity = new JSEntity<T>(driver(), locatorsToBy(locators)).initClass(cl);
+        entity.driver.setBuilder(new JSBuilder(driver(), new SmartBuilderActions()));
         return entity;
     }
     public static JSElement element(String locator) {
-        return new JSElement(DRIVER.get(), NAME_TO_LOCATOR.execute(locator));
+        return new JSElement(driver(), NAME_TO_LOCATOR.execute(locator));
     }
     public static JSElement element(String... locators) {
-        return new JSElement(DRIVER.get(), locatorsToBy(locators));
+        return new JSElement(driver(), locatorsToBy(locators));
     }
     public static <T> JSEntity<T> element(Class<T> cl, String locator) {
-        return new JSEntity<T>(DRIVER.get(), NAME_TO_LOCATOR.execute(locator)).initClass(cl);
+        return new JSEntity<T>(driver(), NAME_TO_LOCATOR.execute(locator)).initClass(cl);
     }
     public static <T> JSEntity<T> element(Class<T> cl, String... locators) {
-        return new JSEntity<T>(DRIVER.get(), locatorsToBy(locators)).initClass(cl);
+        return new JSEntity<T>(driver(), locatorsToBy(locators)).initClass(cl);
     }
     public static JSJson json(String locator) {
-        return new JSJson(DRIVER.get(), NAME_TO_LOCATOR.execute(locator));
+        return new JSJson(driver(), NAME_TO_LOCATOR.execute(locator));
     }
     public static JSJson json(String... locators) {
-        return new JSJson(DRIVER.get(), locatorsToBy(locators));
+        return new JSJson(driver(), locatorsToBy(locators));
     }
 
     public static WebDriver chromeDriver() {
