+++ tests-js-driver/pom.xml
@@ -23,6 +23,12 @@
             <scope>test</scope>
             <version>7.4.0</version>
         </dependency>
+
+        <dependency>
+            <groupId>io.github.bonigarcia</groupId>
+            <artifactId>webdrivermanager</artifactId>
+            <version>4.3.1</version>
+        </dependency>
     </dependencies>
 
     <build>
